% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHist.err.R
\name{AgregHist.err}
\alias{AgregHist.err}
\title{Title  prevision and prevision error}
\usage{
AgregHist.err(xx, grille = aa, nbr = 50, B = 10, dobs, alpha = 1)
}
\arguments{
\item{xx}{data vector}

\item{grille}{grid for density evaluation}

\item{nbr}{number of break sfor histogram}

\item{B}{number of histograms to aggregate}

\item{dobs}{observation}

\item{alpha}{parameters}
}
\value{
prevision and prevision error
}
\description{
Title  prevision and prevision error
}
