% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHist.R
\name{AgregHist}
\alias{AgregHist}
\title{Aggregation of disturbed histograms.}
\usage{
AgregHist(xx, grid, nbr = 50, B = 10, alpha = 1)
}
\arguments{
\item{xx}{data vector for histograms bulding.}

\item{grid}{data vector.this vector makes it possible to provide estimations and estimation errors.}

\item{nbr}{number of breaks  for histogram}

\item{B}{number of histograms to aggregate}

\item{alpha}{disturbance parametter}
}
\value{
Estimation of the values of a density function.
}
\description{
This function aggregates disturbed histograms.Each histogram is disturbed by adding to each of its estimates an observation of a centered normal variable
whose variance is coeficiant "alpha" which is the pertubation parameter.For each histogram a simulation of this normal variable is performed.
}
