% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHist_err.R
\name{AgregHist_err}
\alias{AgregHist_err}
\title{Title  prevision and prevision error according to AgregHist.err}
\usage{
AgregHist_err(xx, grid, nbr = 50, B = 10, dobs, alpha = 1)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid}

\item{nbr}{number of break sfor histogram}

\item{B}{number of histograms to aggregate}

\item{dobs}{observation}

\item{alpha}{parameters}
}
\value{
prevision and prevision error
}
\description{
Title  prevision and prevision error according to AgregHist.err
}
