% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHistunif.err.R
\name{AgregHistunif.err}
\alias{AgregHistunif.err}
\title{Title prevision and prevision error}
\usage{
AgregHistunif.err(xx, grille = aa, nbr = 50, B = 10, dobs)
}
\arguments{
\item{xx}{data vector}

\item{grille}{grid for density evaluation}

\item{nbr}{number of break sfor histogram}

\item{B}{number of histograms to aggregate}

\item{dobs}{observation}
}
\description{
Title prevision and prevision error
}
