% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHistunif.R
\name{AgregHistunif}
\alias{AgregHistunif}
\title{Aggregation of disturbed histograms.}
\usage{
AgregHistunif(xx, grid, nbr = 50, B = 10)
}
\arguments{
\item{xx}{data vector.this vector makes it possible to build histograms}

\item{grid}{data vector.This vector makes it possible to provide estimates and estimation errors.}

\item{nbr}{number of breaks for histogram}

\item{B}{number of histograms to aggregate}
}
\value{
Estimation of the values of a density function .
}
\description{
This function aggregates disturbed histograms.Each histogram is disturbed by adding
small random numbers to the sub-interval edges of the partitions from which they are constructed.
These small random numbers are generated according to a uniform law.
}
