% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHistunif_err.R
\name{AgregHistunif_err}
\alias{AgregHistunif_err}
\title{Aggregation of disturbed histograms and estimation error.}
\usage{
AgregHistunif_err(xx, grid, nbr = 50, B = 10, dobs)
}
\arguments{
\item{xx}{data vector.this vector makes it possible to build histograms}

\item{grid}{data vector.This vector makes it possible to provide estimates and estimation errors.}

\item{nbr}{number of breaks for histogram}

\item{B}{number of histograms to aggregate}

\item{dobs}{density values associated with test sample.}
}
\value{
Estimation of the values of a density function And estimation error.
}
\description{
This function aggregates disturbed histograms an provides estimation error from the aggregation.
This function aggregates disturbed histograms.Each histogram is disturbed by adding small random numbers to
the sub-interval edges of the partitions from which they are constructed.These small random numbers are
generated according to a uniform law.
}
