% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash_err.R
\name{ash_err}
\alias{ash_err}
\title{Title   prevision and prevision error according to ash.err}
\usage{
ash_err(xx, aa, nbr, B = 10, dobs)
}
\arguments{
\item{xx}{data vector}

\item{aa}{grid for density evaluation}

\item{nbr}{number of break sfor histogram}

\item{B}{number of histograms to aggregate}

\item{dobs}{observation}
}
\value{
prevision and prevision error
}
\description{
Title   prevision and prevision error according to ash.err
}
