% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BagHist.err.R
\name{BagHist.err}
\alias{BagHist.err}
\title{Title prevision and prevision error}
\usage{
BagHist.err(xx, grille = aa, B = 10, dobs)
}
\arguments{
\item{xx}{data vector}

\item{grille}{grid for density evaluation}

\item{B}{number of histograms to aggregate}

\item{dobs}{observation}
}
\value{
prevision and prevision error
}
\description{
Title prevision and prevision error
}
