% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BagHist_err.R
\name{BagHist_err}
\alias{BagHist_err}
\title{Density estimation whit BagHist_err}
\usage{
BagHist_err(xx, grid, B = 10, dobs)
}
\arguments{
\item{xx}{data vector for histograms construction .}

\item{grid}{grid for density evaluation}

\item{B}{number of histograms to aggregate}

\item{dobs}{density values associated whit test sample}
}
\value{
estimation and estimation error.
}
\description{
This function builds histograms from the bootstrap samples.
Then averages  the estimates provided  by these histograms and computes these estimation error.
}
