% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BagHistfp.err.R
\name{BagHistfp.err}
\alias{BagHistfp.err}
\title{Title  prevision and prevision error}
\usage{
BagHistfp.err(xx, grille = aa, B = 10, dobs)
}
\arguments{
\item{xx}{data vector}

\item{grille}{grid for density evaluation}

\item{B}{number of histograms to aggregate}

\item{dobs}{observation}
}
\value{
prevision and prevision error
}
\description{
Title  prevision and prevision error
}
