% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BagHistfp.R
\name{BagHistfp}
\alias{BagHistfp}
\title{Title prevision and prevision acording to BagHistfp}
\usage{
BagHistfp(xx, grille = aa, B = 10)
}
\arguments{
\item{xx}{data vector}

\item{grille}{grid for density evaluation}

\item{B}{number of histograms to aggregate}
}
\value{
prevision and prevision
}
\description{
Title prevision and prevision acording to BagHistfp
}
