% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BagHistfp_err.R
\name{BagHistfp_err}
\alias{BagHistfp_err}
\title{Title  prevision and prevision error according to BagHistfp.err}
\usage{
BagHistfp_err(xx, grid, B = 10, dobs)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid}

\item{B}{number of histograms to aggregate}

\item{dobs}{observation}
}
\value{
prevision and prevision error
}
\description{
Title  prevision and prevision error according to BagHistfp.err
}
