% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagkde.R
\name{Bagkde}
\alias{Bagkde}
\title{Density estimation with Bagkde}
\usage{
Bagkde(xx, grid, dobs, B = 10)
}
\arguments{
\item{xx}{data vector  for estimator bulding}

\item{grid}{grid for density evaluation}

\item{dobs}{density values associated whit test sample}

\item{B}{number of histograms to aggregate}
}
\value{
estimations
}
\description{
This function builds Kernel Densities estimators from the bootstrap samples.
Then averages  the estimates provided  by these estimators.
}
