% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BagKDE_err.R
\name{BagKDE_err}
\alias{BagKDE_err}
\title{Title prevision and prevision error  according to BagKDE.err}
\usage{
BagKDE_err(xx, grid, B = 10, dobs)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid for density evaluation}

\item{B}{number of histograms to aggregate}

\item{dobs}{obervation}
}
\value{
prevision and prevision error
}
\description{
Title prevision and prevision error  according to BagKDE.err
}
