% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseboot.R
\name{baseboot}
\alias{baseboot}
\title{Variability band}
\usage{
baseboot(x, pointwise, bunch, conf = c(0.05, 0.95), plot = FALSE)
}
\arguments{
\item{x}{data vector}

\item{pointwise}{a vector with the pointwise density estimation}

\item{bunch}{bootstrap replicated of the estimation density (matrix)}

\item{conf}{confidence quantiles}

\item{plot}{logical, make a validation plot}
}
\value{
matrix with lower and upper bounds of the variability band
}
\description{
Computation of a variability band using base bootstrap
}
