% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error}
\alias{error}
\title{Estimation error}
\usage{
error(obs, prev)
}
\arguments{
\item{obs}{values associated with test sample.}

\item{prev}{prevision}
}
\value{
quadratic error ,Error log likelihood
}
\description{
this function provides estimation  using quadratic error or Error log likelihood .
}
