% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error2.R
\name{error2}
\alias{error2}
\title{Estimation error}
\usage{
error2(obs, prev)
}
\arguments{
\item{obs}{sample of observations}

\item{prev}{prevision}
}
\value{
Quadratic error ,Error log likelihood or  Hellinger dissimilarity error
}
\description{
This function provides the error made by an estimator
}
