% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{gendata}
\alias{gendata}
\title{Title  the Training set, the Test set}
\usage{
gendata(nummodel = 1, n = 100)
}
\arguments{
\item{nummodel}{number of the model}

\item{n}{sample size}
}
\value{
the training s,et, the test set
}
\description{
Title  the Training set, the Test set
}
