% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{KDE}
\alias{KDE}
\title{Title  ????}
\usage{
KDE(dlearn, dtest, w = NULL, h = 0.1)
}
\arguments{
\item{dlearn}{learning sample}

\item{dtest}{test sample}

\item{w}{?????}

\item{h}{????}
}
\value{
?????
}
\description{
Title  ????
}
