% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybreaks.R
\name{mybreaks}
\alias{mybreaks}
\title{breaks}
\usage{
mybreaks(x = rnorm(100), nbr)
}
\arguments{
\item{x}{data vector}

\item{nbr}{number of breaks for an estimator}
}
\value{
a breaks
}
\description{
This funtion  provides breaks for a density estimator
}
