% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onekdeucv.R
\name{onekdeucv}
\alias{onekdeucv}
\title{Title  kde estimation}
\usage{
onekdeucv(xx, grid)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid}
}
\value{
estimations
}
\description{
Title  kde estimation
}
