% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_hist.R
\name{predict_hist}
\alias{predict_hist}
\title{Prediction whith histogram}
\usage{
predict_hist(hh, x)
}
\arguments{
\item{hh}{This function estimates values associated with a sample of data using the histogram.}

\item{x}{sample of observations}
}
\value{
prediction
}
\description{
this function
}
