% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rash.err.R
\name{rash.err}
\alias{rash.err}
\title{Title  prevision and prevision error}
\usage{
rash.err(xx, grille = aa, nbr = 50, B = 10, dobs)
}
\arguments{
\item{xx}{data vector}

\item{grille}{grid for density evaluation}

\item{nbr}{number of break sfor histogram}

\item{B}{number of histograms to aggregate}

\item{dobs}{observation}
}
\value{
prevision and prevision error
}
\description{
Title  prevision and prevision error
}
