% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube_fp.R
\name{tube_fp}
\alias{tube_fp}
\title{Construction of a variability band with frquency polygons.}
\usage{
tube_fp(data, nbr = 50, B = 10, conf = c(0.05, 0.95))
}
\arguments{
\item{data}{data of test sample and learning  (data$learn and data$test).
test sample must be sorted  (sort(data$test)).}

\item{nbr}{number of sub-intervals of the initial partition}

\item{B}{Number of bootstrap replications}

\item{conf}{vector of the confidence threshold}
}
\value{
The vector of values of the aggregated version of a frquency polygon, a matrix of predictions
        of frquency polygon, matrix of the lower edge and the upper edge of a variability band.
}
\description{
At each step, this function takes a new set of data, builds a
             frquency polygon,predicts and aggregates.
}
