% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube_KDE.R
\name{tube_KDE}
\alias{tube_KDE}
\title{Construction of a variability band with kernel density estimators.}
\usage{
tube_KDE(data, B = 10, conf = c(0.05, 0.95))
}
\arguments{
\item{data}{ata of test sample and learning  (data$learn and data$test).
test sample must be sorted  (sort(data$test)).}

\item{B}{Number of bootstrap replications}

\item{conf}{vector of the confidence threshold}
}
\value{
The vector of values of the aggregated version of a kernel density estimator, a matrix of predictions
        of kernel density estimator, matrix of the lower edge and the upper edge of a variability band.
}
\description{
At each step, this function takes a new set of data, builds
             a kernel density estimator, predicts and aggregates.
}
