% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube_KDEsm.R
\name{tube_KDEsm}
\alias{tube_KDEsm}
\title{Computes the confidence band using smooth kernel density estimators}
\usage{
tube_KDEsm(data)
}
\arguments{
\item{data}{ata of test sample and learning  (data$learn and data$test).
test sample must be sorted  (sort(data$test)).}
}
\value{
the lower edge and the upper edge of a confinement band
        and the aggregated version of the initial kernel density estimator.
}
\description{
this function builds a confidence band from the kernel density estimators
}
