% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube_wass.R
\name{tube_wass}
\alias{tube_wass}
\title{confidence band with histograms}
\usage{
tube_wass(data, alpha = 0.1, nbr = 50, plot = FALSE, ...)
}
\arguments{
\item{data}{ata of test sample and learning  (data$learn and data$test).
test sample must be sorted  (sort(data$test)).}

\item{alpha}{parameter for qnorm}

\item{nbr}{number of breaks for histogram}

\item{plot}{TRUE or FALSE, to plot confidence band}

\item{...}{additional parameters for the histogram}
}
\value{
the lower edge and the upper edge of a confinement band
        and the aggregated version of the initial histogram.
}
\description{
this function builds a confidence band from the histograms
}
