% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHist.R
\name{AgregHist}
\alias{AgregHist}
\title{Density estimation with AgregHist.}
\usage{
AgregHist(xx, grid, nbr = 50, B = 10, alpha = 1)
}
\arguments{
\item{xx}{data vector for histograms bulding.}

\item{grid}{grid for density evaluation.}

\item{nbr}{number of breaks  for histogram}

\item{B}{number of histograms to aggregate}

\item{alpha}{disturbance parametter}
}
\value{
Estimation.
}
\description{
This function aggregates disturbed histograms.Each histogram is disturbed by adding
small random numbers to the sub-interval edges of the partitions from which they are constructed.
}
