% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHist.R
\name{AgregHist}
\alias{AgregHist}
\title{Title  prevision and prevision error according to AgregHist}
\usage{
AgregHist(xx, grid, nbr = 50, B = 10, alpha = 1)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid}

\item{nbr}{number of break sfor histogram}

\item{B}{number of histograms to aggregate}

\item{alpha}{disturbance parametter}
}
\value{
prevision
}
\description{
Title  prevision and prevision error according to AgregHist
}
