% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHist.R
\name{AgregHist}
\alias{AgregHist}
\title{Title  ???????}
\usage{
AgregHist(xx, grille = aa, nbr = 50, B = 10, alpha = 1)
}
\arguments{
\item{xx}{data vector}

\item{grille}{grid for density evaluation}

\item{B}{number of histograms to aggregate}

\item{alpha}{disturbance parametter}
}
\value{
????
}
\description{
Title  ???????
}
