% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHist_err.R
\name{AgregHist_err}
\alias{AgregHist_err}
\title{Aggregation of disturbed histograms and estimation error.}
\usage{
AgregHist_err(xx, grid, nbr = 50, B = 10, dobs, alpha = 1)
}
\arguments{
\item{xx}{data vector . this vector makes it possible to build histograms.}

\item{grid}{data vector.this vector makes it possible to provide estimates and estimation errors.}

\item{nbr}{number of breaks for histogram}

\item{B}{number of histograms to aggregate}

\item{dobs}{density values associated with test sample.}

\item{alpha}{pertubation parametter}
}
\value{
Estimation of the values of a density function And estimation error.
}
\description{
This function aggregates disturbed histograms an provides estimation error from the aggregation.
Each histogram is disturbed by adding to each of its estimates an observation of a centered normal variable whose
variance is coeficiant "alpha" which is the pertubation parameter.
For each histogram a simulation of the normal variable is performed.
}
