% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHistunif.R
\name{AgregHistunif}
\alias{AgregHistunif}
\title{Title  ???????}
\usage{
AgregHistunif(xx, grille = aa, nbr = 50, B = 10)
}
\arguments{
\item{xx}{data vector}

\item{grille}{grid for density evaluation}

\item{nbr}{number of break sfor histogram}

\item{B}{number of histograms to aggregate}
}
\value{
??????
}
\description{
Title  ???????
}
