% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHistunif.R
\name{AgregHistunif}
\alias{AgregHistunif}
\title{Title  prevision and prevision error according to AgregHistunif}
\usage{
AgregHistunif(xx, grid, nbr = 50, B = 10)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid}

\item{nbr}{number of break sfor histogram}

\item{B}{number of histograms to aggregate}
}
\value{
prevision
}
\description{
Title  prevision and prevision error according to AgregHistunif
}
