% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHistunif.R
\name{AgregHistunif}
\alias{AgregHistunif}
\title{Aggregation of disturbed histograms.}
\usage{
AgregHistunif(xx, grid, nbr = 50, B = 10)
}
\arguments{
\item{xx}{data vector.this vector makes it possible to build histograms}

\item{grid}{data vector.This vector makes it possible to provide estimates and estimation errors.}

\item{nbr}{number of breaks for histogram}

\item{B}{number of histograms to aggregate}
}
\value{
Estimation of the values of a density function .
}
\description{
This function aggregated histograms disturbed.each histogram is disturbed with a uniform variable.
Thus for each histogram a simulation of this uniform variable is performed.
}
