% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgregHistunif_err.R
\name{AgregHistunif_err}
\alias{AgregHistunif_err}
\title{Title prevision and prevision error  according to AgregHistunif.err}
\usage{
AgregHistunif_err(xx, grid, nbr = 50, B = 10, dobs)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid}

\item{nbr}{number of break sfor histogram}

\item{B}{number of histograms to aggregate}

\item{dobs}{observation}
}
\value{
prevision and prevision error
}
\description{
Title prevision and prevision error  according to AgregHistunif.err
}
