% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash_err.R
\name{ash_err}
\alias{ash_err}
\title{Density estimation whit ash_err}
\usage{
ash_err(xx, aa, nbr, B = 10, dobs)
}
\arguments{
\item{xx}{data vector for histograms construction .}

\item{aa}{grid for density evaluation}

\item{nbr}{number of break sfor histogram}

\item{B}{number of histograms to aggregate}

\item{dobs}{density values associated whit test sample}
}
\value{
prevision and prevision error
}
\description{
this function builds several histograms on partions obtained
             by non-randomly translating an initial partition.
}
