% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BagHist.R
\name{BagHist}
\alias{BagHist}
\title{Density Estimation with Bagged Histograms (BagHist)}
\usage{
BagHist(xx, grid, B = 10)
}
\arguments{
\item{xx}{numeric vector.}

\item{grid}{numeric vector used for evaluation}

\item{B}{number of bootstrap samples.}
}
\value{
numeric vector with the density estimated over the given grid using BagHist

estimation
}
\description{
Estimate the density of a univariate data vector using BagHist. First,
obtain bootstrap samples of the input vector data. For each sample, a histogram is
computed using an optimized number of break points. The BagHist estimation is the mean of the individual histograms.
}
\examples{
x <- rnorm(100)
BagHist(x, seq(-6, 6, length.out = 100), B = 50)
}
