% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BagHist.R
\name{BagHist}
\alias{BagHist}
\title{Title prevision and prevision error}
\usage{
BagHist(xx, grid, B = 10)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid}

\item{B}{number of histograms to aggregate}
}
\value{
prevision and prevision error
}
\description{
Title prevision and prevision error
}
