% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BagHist_err.R
\name{BagHist_err}
\alias{BagHist_err}
\title{Title prevision and prevision error according to BagHist.err}
\usage{
BagHist_err(xx, grid, B = 10, dobs)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid}

\item{B}{number of histograms to aggregate}

\item{dobs}{observation}
}
\value{
prevision and prevision error
}
\description{
Title prevision and prevision error according to BagHist.err
}
