% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BagHistfp.R
\name{BagHistfp}
\alias{BagHistfp}
\title{Density Estimation with Bagged Frequency Polygons (BagHistFP)}
\usage{
BagHistfp(xx, grid, B = 10)
}
\arguments{
\item{xx}{numeric vector.}

\item{grid}{numeric vector used for evaluation}

\item{B}{number of bootstrap samples.}
}
\value{
numeric vector with the density estimated over the given grid using BagHistFP
}
\description{
Estimate the density of a univariate data vector using BagHistFP. First,
obtain bootstrap samples of the input vector data. For each sample, a histogram is
computed using an optimized number of break points. On top of it, a frequency polygon
is obtained. The BagHistFP estimation is the mean of the individual frequency polygons.
}
\examples{
x <- rnorm(100)
BagHistfp(x, seq(-6, 6, length.out = 100), B = 50)
}
