% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BagHistfp.R
\name{BagHistfp}
\alias{BagHistfp}
\title{Density estimation whit BagHistfp}
\usage{
BagHistfp(xx, grid, B = 10)
}
\arguments{
\item{xx}{data vector for Frequency Folygone construction .}

\item{grid}{grid for density evaluation}

\item{B}{number of Frequency Folygone to aggregate}
}
\value{
estimation
}
\description{
This function builds "Frequency Folygones"  from the bootstrap samples.
Then averages  the estimates provided  by these estimators .
}
