% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagkde.R
\name{Bagkde}
\alias{Bagkde}
\title{Title Prevision acording to Bagkde}
\usage{
Bagkde(xx, grid, dobs, B = 10)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid}

\item{dobs}{observation}

\item{B}{number of histograms to aggregate}
}
\value{
Prevision
}
\description{
Title Prevision acording to Bagkde
}
