% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagkde.R
\name{Bagkde}
\alias{Bagkde}
\title{TitlePrevision acording to Bagkde}
\usage{
Bagkde(xx, grille = aa, dobs, B = 10)
}
\arguments{
\item{xx}{data vector}

\item{grille}{grid for density evaluation}

\item{dobs}{observation}

\item{B}{number of histograms to aggregate}
}
\value{
Prevision
}
\description{
TitlePrevision acording to Bagkde
}
