% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BagKDE_err.R
\name{BagKDE_err}
\alias{BagKDE_err}
\title{Density estimation with Bagkde_err}
\usage{
BagKDE_err(xx, grid, B = 10, dobs)
}
\arguments{
\item{xx}{data vector  for estimator bulding}

\item{grid}{grid for density evaluation}

\item{B}{number of histograms to aggregate}

\item{dobs}{density values associated whit test sample}
}
\value{
prevision and prevision error
}
\description{
This function builds Kernel Densities estimators from the bootstrap samples.
Then averages  the estimates provided  by these estimators  and computes their estimation error.
}
