% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseboot.R
\name{baseboot}
\alias{baseboot}
\title{variability band}
\usage{
baseboot(x, punctual, bunch, conf = c(0.05, 0.95), plot = FALSE)
}
\arguments{
\item{x}{data vector}

\item{punctual}{aggregation data vector}

\item{bunch}{estimations matrice}

\item{conf}{confidence threshold vector}

\item{plot}{plot validation}
}
\value{
the lower edge and the upper edge of a variability band
}
\description{
This function computes a variability band from a set of vectors
}
