% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{gendata}
\alias{gendata}
\title{Data generation whith gendata  function .}
\usage{
gendata(nummodel = 1, n = 100)
}
\arguments{
\item{nummodel}{number to choose a dristribution low.}

\item{n}{sample size .}
}
\value{
the training set, the test set and density values .
}
\description{
This function provides the Training set, the Test set and density values
from a distribution law.
}
\examples{
gendata(10,100)

}
