% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{KDE}
\alias{KDE}
\title{Title  prevision according to KDE}
\usage{
KDE(dlearn, dtest, w = NULL, h = 0.1)
}
\arguments{
\item{dlearn}{learning sample}

\item{dtest}{test sample}

\item{w}{????}

\item{h}{step}
}
\value{
prvision
}
\description{
Title  prevision according to KDE
}
