% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDEtr.R
\name{KDEtr}
\alias{KDEtr}
\title{Title ???????}
\usage{
KDEtr(dlearn, dtest, w = NULL, h = 0.1)
}
\arguments{
\item{dlearn}{learning saple}

\item{dtest}{test sample}

\item{w}{?????}

\item{h}{?????}
}
\value{
?????
}
\description{
Title ???????
}
