% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDEtr.R
\name{KDEtr}
\alias{KDEtr}
\title{estimation according to KDEtr}
\usage{
KDEtr(dlearn, dtest, w = NULL, h = 0.1)
}
\arguments{
\item{dlearn}{learning sample}

\item{dtest}{test sample}

\item{w}{a vector of 1/n , where n is the length learning sample .}

\item{h}{step}
}
\value{
Estimation
}
\description{
This function makes density estimation whith Kernel Density Estimator using
             triangular kernel.
}
