% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybreaks.R
\name{mybreaks}
\alias{mybreaks}
\title{Title breaks}
\usage{
mybreaks(x = rnorm(100), nbr)
}
\arguments{
\item{x}{data vector}

\item{nbr}{number of break sfor histogram}
}
\value{
a breaks
}
\description{
Title breaks
}
