% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onekdeucv.R
\name{onekdeucv}
\alias{onekdeucv}
\title{kde estimation}
\usage{
onekdeucv(xx, grid)
}
\arguments{
\item{xx}{data vector  for estimator bulding}

\item{grid}{grid for density evaluation}
}
\value{
estimations
}
\description{
this function estimates a density using kde
}
