% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onekdeucv.R
\name{onekdeucv}
\alias{onekdeucv}
\title{Title ??}
\usage{
onekdeucv(xx, grille = aa)
}
\arguments{
\item{xx}{data vector}

\item{grille}{grille    grid for density evaluation}
}
\value{
??
}
\description{
Title ??
}
