% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_hist.R
\name{predict_hist}
\alias{predict_hist}
\title{Title Prediction according to the histogram}
\usage{
predict_hist(hh, x)
}
\arguments{
\item{hh}{????}

\item{x}{?????}
}
\value{
prediction
}
\description{
Title Prediction according to the histogram
}
