% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rash.R
\name{rash}
\alias{rash}
\title{Density estimation rash.}
\usage{
rash(xx, grid, nbr = 50, B = 10)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid for density evaluation}

\item{nbr}{number of breaks for histogram}

\item{B}{number of histograms to aggregate}
}
\value{
Prevision
}
\description{
This function builds several histograms by disturbing the partition
             of the histogram with a Gaussian noise and then aggregates them.
}
