% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rash.R
\name{rash}
\alias{rash}
\title{Title  Prevision acording to rash}
\usage{
rash(xx, grid, nbr = 50, B = 10)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid}

\item{nbr}{number of break sfor histogram}

\item{B}{number of histograms to aggregate}
}
\value{
Prevision
}
\description{
Title  Prevision acording to rash
}
