% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rash_err.R
\name{rash_err}
\alias{rash_err}
\title{Title  prevision and prevision error}
\usage{
rash_err(xx, grid, nbr = 50, B = 10, dobs)
}
\arguments{
\item{xx}{data vector}

\item{grid}{grid}

\item{nbr}{number of break sfor histogram}

\item{B}{number of histograms to aggregate}

\item{dobs}{observation}
}
\value{
prevision and prevision error
}
\description{
Title  prevision and prevision error
}
